Initial rework.
