--- .rspec
@@ -1,4 +1,0 @@
---color 
---format documentation
---warnings
---require spec_helper
