--- .travis.yml
@@ -1,21 +1,0 @@
-language: ruby
-cache: bundler
-
-addons:
-  apt:
-    packages:
-    - bind9
-
-matrix:
-  include:
-    - rvm: 2.3
-    - rvm: 2.4
-    - rvm: 2.5
-    - rvm: 2.6
-    - rvm: 2.7
-    - rvm: jruby-head
-      env: JRUBY_OPTS="--debug -X+O"
-    - rvm: ruby-head
-  allow_failures:
-    - rvm: ruby-head
-    - rvm: jruby-head
