--- Gemfile
@@ -1,13 +1,0 @@
-source 'https://rubygems.org'
-
-gemspec
-
-group :development do
-	gem "process-daemon"
-	gem "nio4r"
-end
-
-group :test do
-	gem 'simplecov'
-	gem 'coveralls', require: false
-end
