--- Rakefile
@@ -1,6 +1,0 @@
-require "bundler/gem_tasks"
-require "rspec/core/rake_task"
-
-RSpec::Core::RakeTask.new
-
-task :default => :spec
