+++ bin/rubydns-check
@@ -1,4 +1,5 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
 # Copyright, 2009, 2012, by Samuel G. D. Williams. <http://www.codeotaku.com>
 # 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -28,59 +29,59 @@
 # Check data against new DNS server
 # rd-dns-check -s 10.0.0.36 -d mydomain.com. -c old-dns.yml
 
-require 'yaml'
-require 'optparse'
-require 'set'
+require "yaml"
+require "optparse"
+require "set"
 
 class DNSRecord
 	def initialize(arr)
 		@record = arr
 		normalize
 	end
-  
+		
 	def normalize
 		@record[0] = @record[0].downcase
 		@record[1] = @record[1].upcase
 		@record[2] = @record[2].upcase
 		@record[3] = @record[3].downcase
 	end
-  
+		
 	def hostname
 		@record[0]
 	end
-  
+		
 	def klass
 		@record[1]
 	end
-  
+		
 	def type
 		@record[2]
 	end
-  
+		
 	def value
 		@record[3]
 	end
-  
+		
 	def is_address?
 		["A", "AAAA"].include?(type)
 	end
-  
+		
 	def is_cname?
 		return type == "CNAME"
 	end
-  
+		
 	def to_s
 		"#{hostname.ljust(50)} #{klass.rjust(4)} #{type.rjust(5)} #{value}"
 	end
-  
+		
 	def key
 		"#{hostname}:#{klass}:#{type}".downcase
 	end
-  
+		
 	def to_a
 		@record
 	end
-  
+		
 	def == other
 		return @record == other.to_a
 	end
@@ -94,7 +95,7 @@ def dig(dns_server, cmd, exclude = ["TXT", "HINFO", "SOA", "NS"])
 			r = line.chomp.split(/\s/, 4)
 			
 			next if exclude.include?(r[2])
-      
+						
 			records << DNSRecord.new(r)
 		end
 	end
@@ -137,7 +138,7 @@ end
 def resolve_addresses(records)
 	addresses = {}
 	cnames = {}
-  
+		
 	# Extract all hostname -> ip address mappings
 	records.each do |r|
 		if r.is_address?
@@ -146,17 +147,17 @@ def resolve_addresses(records)
 			cnames[r.hostname] = r
 		end
 	end
-  
+		
 	cnames.each do |hostname, r|
 		q = r
 		trail = []
 		failed = false
-    
+				
 		# Keep track of CNAME records to avoid loops
 		while q.is_cname?
 			trail << q
 			q = cnames[q.value] || addresses[q.value]
-      
+						
 			# Q could be nil at this point, which means there was no address record
 			# Q could be already part of the trail, which means there was a loop
 			if q == nil || trail.include?(q)
@@ -164,7 +165,7 @@ def resolve_addresses(records)
 				break
 			end
 		end
-    
+				
 		if failed
 			q = trail.last
 			puts "*** Warning: CNAME record #{hostname} does not point to actual address!"
@@ -172,24 +173,24 @@ def resolve_addresses(records)
 				puts idx.to_s.rjust(10) + ": " + r.to_s
 			end
 		end
-    
+				
 		addresses[r.hostname] = q
 	end
-  
+		
 	return addresses, cnames
 end
 
 def check_reverse(records, dns_server)
 	errors = 0
 	okay = []
-  
+		
 	puts "[ Checking Reverse Lookups ]".center(72, "=")
-  
+		
 	records.each do |r|
 		next unless r.is_address?
-    
+				
 		sr = resolve_address(dns_server, r.value)
-    
+				
 		if sr == nil
 			puts "*** Could not resolve host"
 			puts "".rjust(12) + r.to_s
@@ -203,7 +204,7 @@ def check_reverse(records, dns_server)
 			okay << [r, sr]
 		end
 	end
-  
+		
 	print_summary(records, errors, okay) do |r|
 		puts "Primary:".rjust(12) + r[0].to_s
 		puts "Secondary:".rjust(12) + r[1].to_s
@@ -217,12 +218,12 @@ def ping_records(records)
 	okay = []
 
 	puts "[ Pinging Records ]".center(72, "=")
-  
+		
 	addresses.each do |hostname, r|
 		ping = "ping -c 5 -t 5 -i 1 -o #{r.value} > /dev/null"
-    
+				
 		system(ping)
-    
+				
 		if $?.exitstatus == 0
 			okay << r
 		else
@@ -231,31 +232,31 @@ def ping_records(records)
 			errors += 1
 		end
 	end
-  
+		
 	print_summary(records, errors, okay)
 end
 
 def query_records(primary, secondary_server)
 	addresses, cnames = resolve_addresses(primary)
-  
+		
 	okay = []
 	errors = 0
-  
+		
 	primary.each do |r|
 		sr = resolve_hostname(secondary_server, r.hostname)
-    
+				
 		if sr == nil
 			puts "*** Could not resolve hostname #{r.hostname.dump}"
 			puts "Primary: ".rjust(12) + r.to_s
-      
+						
 			rsr = resolve_address(secondary_server, (addresses[r.value] || r).value)
 			puts "Address: ".rjust(12) + rsr.to_s if rsr
-      
+						
 			errors += 1
 		elsif sr.value != r.value
 			ra = addresses[r.value] if r.is_cname?
 			sra = addresses[sr.value] if sr.is_cname?
-      
+						
 			if (sra || sr).value != (ra || r).value
 				puts "*** IP Address does not match"
 				puts "Primary: ".rjust(12) + r.to_s
@@ -268,7 +269,7 @@ def query_records(primary, secondary_server)
 			okay << r
 		end
 	end
-  
+		
 	print_summary(primary, errors, okay)
 end
 
@@ -276,7 +277,7 @@ def check_records(primary, secondary)
 	s = {}
 	okay = []
 	errors = 0
-  
+		
 	secondary.each do |r|
 		s[r.key] = r
 	end
@@ -285,7 +286,7 @@ def check_records(primary, secondary)
 
 	primary.each do |r|
 		sr = s[r.key]
-    
+				
 		if sr == nil
 			puts "*** Could not find record"
 			puts "Primary: ".rjust(12) + r.to_s
@@ -299,7 +300,7 @@ def check_records(primary, secondary)
 			okay << r
 		end
 	end
-  
+		
 	print_summary(primary, errors, okay)
 end
 
@@ -310,65 +311,65 @@ OPTIONS = {
 
 ARGV.options do |o|
 	script_name = File.basename($0)
-  
-	o.set_summary_indent('  ')
+		
+	o.set_summary_indent("  ")
 	o.banner = "Usage: #{script_name} [options]"
 	o.define_head "This script is designed to test and check DNS servers."
-  
+		
 	o.on("-s ns.my.domain.", "--server ns.my.domain.", String, "The DNS server to query.") { |host| OPTIONS[:DNSServer] = host }
 	o.on("-d my.domain.", "--domain my.domain.", String, "The DNS zone to transfer/test.") { |host| OPTIONS[:DNSRoot] = host }
-  
+		
 	o.on("-f output.yml", "--fetch output.yml", String, "Pull down a list of hosts. Filters TXT and HINFO records. DNS transfers must be enabled.") { |f|
 		records = retrieve_records(OPTIONS[:DNSServer], OPTIONS[:DNSRoot])
-    
+				
 		output = (f ? File.open(f, "w") : STDOUT)
-    
+				
 		output.write(YAML::dump(records))
-    
+				
 		puts "#{records.size} record(s) retrieved."
 	}
-  
+		
 	o.on("-c input.yml", "--check input.yml", String, "Check that the DNS server returns results as specified by the file.") { |f|
 		input = (f ? File.open(f) : STDIN)
-    
+				
 		master_records = YAML::load(input.read)
 		secondary_records = retrieve_records(OPTIONS[:DNSServer], OPTIONS[:DNSRoot])
-    
+				
 		check_records(master_records, secondary_records)
 	}
-  
+		
 	o.on("-q input.yml", "--query input.yml", String, "Query the remote DNS server with all hostnames in the given file, and checks the IP addresses are consistent.") { |f|
 		input = (f ? File.open(f) : STDIN)
-    
+				
 		master_records = YAML::load(input.read)
-    
+				
 		query_records(master_records, OPTIONS[:DNSServer])
 	}
-  
+		
 	o.on("-p input.yml", "--ping input.yml", String, "Ping all hosts to check if they are available or not.") { |f|
 		input = (f ? File.open(f) : STDIN)
-    
+				
 		master_records = YAML::load(input.read)
-    
+				
 		ping_records(master_records)
 	}
-  
+		
 	o.on("-r input.yml", "--reverse input.yml", String, "Check that all address records have appropriate reverse entries.") { |f|
 		input = (f ? File.open(f) : STDIN)
-    
+				
 		master_records = YAML::load(input.read)
-    
+				
 		check_reverse(master_records, OPTIONS[:DNSServer])
 	}
-  
+		
 	o.separator ""
 	o.separator "Help and Copyright information"
-  
+		
 	o.on_tail("--copy", "Display copyright information") {
 		puts "#{script_name}. Copyright (c) 2009, 2011 Samuel Williams. Released under the MIT license."
 		puts "See http://www.oriontransfer.co.nz/ for more information."
 		exit
 	}
-  
+		
 	o.on_tail("-h", "--help", "Show this help message.") { puts o; exit }
 end.parse!
