+++ examples/gems.locked
@@ -1,0 +1,62 @@
+PATH
+  remote: ..
+  specs:
+    rubydns (2.0.2)
+      async-dns (~> 1.0)
+
+GEM
+  remote: https://rubygems.org/
+  specs:
+    async (2.1.0)
+      console (~> 1.10)
+      io-event (~> 1.0.0)
+      timers (~> 4.1)
+    async-dns (1.3.0)
+      async-io (~> 1.15)
+    async-http (0.59.2)
+      async (>= 1.25)
+      async-io (>= 1.28)
+      async-pool (>= 0.2)
+      protocol-http (~> 0.23.1)
+      protocol-http1 (~> 0.14.0)
+      protocol-http2 (~> 0.14.0)
+      traces (>= 0.4.0)
+    async-io (1.34.0)
+      async
+    async-pool (0.3.12)
+      async (>= 1.25)
+    console (1.15.3)
+      fiber-local
+    fiber-local (1.0.0)
+    geoip (1.6.4)
+    io-event (1.0.9)
+    nokogiri (1.13.8-arm64-darwin)
+      racc (~> 1.4)
+    process-daemon (1.0.1)
+      rainbow (~> 2.0)
+    protocol-hpack (1.4.2)
+    protocol-http (0.23.12)
+    protocol-http1 (0.14.6)
+      protocol-http (~> 0.22)
+    protocol-http2 (0.14.2)
+      protocol-hpack (~> 1.4)
+      protocol-http (~> 0.18)
+    racc (1.6.0)
+    rainbow (2.2.2)
+      rake
+    rake (13.0.6)
+    timers (4.3.4)
+    traces (0.7.0)
+
+PLATFORMS
+  arm64-darwin-21
+
+DEPENDENCIES
+  async-http
+  geoip
+  nokogiri
+  process-daemon
+  rubydns!
+
+BUNDLED WITH
+   2.3.10
