+++ examples/gems.rb
@@ -1,9 +1,15 @@
-source 'https://rubygems.org'
+# frozen_string_literal: true
 
-gem 'rubydns', path: '../'
+# Released under the MIT License.
+# Copyright, 2014-2022, by Samuel Williams.
+# Copyright, 2014, by Peter M. Goldstein.
+
+source "https://rubygems.org"
+
+gem "rubydns", path: "../"
 gem "process-daemon"
-gem 'nokogiri'
+gem "nokogiri"
 
-gem 'async-http'
+gem "async-http"
 
-gem 'geoip'
+gem "geoip"
