+++ examples/soa-dns.rb
@@ -1,4 +1,8 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
+
+# Released under the MIT License.
+# Copyright, 2017, by Samuel Williams.
 
 # Copyright, 2009, 2012, by Samuel G. D. Williams. <http://www.codeotaku.com>
 #
@@ -20,25 +24,25 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-require 'rubygems'
-require 'rubydns'
+require "rubygems"
+require "rubydns"
 
 $R = Resolv::DNS.new
 Name = Resolv::DNS::Name
 IN = Resolv::DNS::Resource::IN
 
-RubyDNS.run_server([[:udp, '0.0.0.0', 5400]]) do
+RubyDNS.run_server([[:udp, "0.0.0.0", 5400]]) do
 	# SOA Record
 	#   dig @localhost -p 5400 SOA mydomain.org
-	match('mydomain.org', IN::SOA) do |transaction|
+	match("mydomain.org", IN::SOA) do |transaction|
 		#
 		# For more details about these headers please see:
 		#   http://www.ripe.net/ripe/docs/ripe-203.html
 		#
 		
 		transaction.respond!(
-			Name.create('ns.mydomain.org.'),    # Master Name
-			Name.create('admin.mydomain.org.'), # Responsible Name
+			Name.create("ns.mydomain.org."),    # Master Name
+			Name.create("admin.mydomain.org."), # Responsible Name
 			File.mtime(__FILE__).to_i,          # Serial Number
 			1200,                               # Refresh Time
 			900,                                # Retry Time
@@ -51,27 +55,27 @@ RubyDNS.run_server([[:udp, '0.0.0.0', 5400]]) do
 
 	# Default NS record
 	#   dig @localhost -p 5400 mydomain.org NS
-	match('mydomain.org', IN::NS) do |transaction|
-		transaction.respond!(Name.create('ns.mydomain.org.'))
+	match("mydomain.org", IN::NS) do |transaction|
+		transaction.respond!(Name.create("ns.mydomain.org."))
 	end
 
 	# For this exact address record, return an IP address
 	#   dig @localhost -p 5400 CNAME bob.mydomain.org
 	match(/([^.]+).mydomain.org/, IN::CNAME) do |transaction|
-		transaction.respond!(Name.create('www.mydomain.org'))
-		transaction.append!('www.mydomain.org', IN::A)
+		transaction.respond!(Name.create("www.mydomain.org"))
+		transaction.append!("www.mydomain.org", IN::A)
 	end
 
-	match('80.0.0.10.in-addr.arpa', IN::PTR) do |transaction|
-		transaction.respond!(Name.create('www.mydomain.org.'))
+	match("80.0.0.10.in-addr.arpa", IN::PTR) do |transaction|
+		transaction.respond!(Name.create("www.mydomain.org."))
 	end
 
-	match('www.mydomain.org', IN::A) do |transaction|
-		transaction.respond!('10.0.0.80')
+	match("www.mydomain.org", IN::A) do |transaction|
+		transaction.respond!("10.0.0.80")
 	end
 
-	match('ns.mydomain.org', IN::A) do |transaction|
-		transaction.respond!('10.0.0.10')
+	match("ns.mydomain.org", IN::A) do |transaction|
+		transaction.respond!("10.0.0.10")
 	end
 
 	# Default DNS handler
