+++ license.md
@@ -1,0 +1,39 @@
+# MIT License
+
+Copyright, 2009-2022, by Samuel Williams.  
+Copyright, 2011, by Genki Sugawara.  
+Copyright, 2012, by Satoshi Takada.  
+Copyright, 2013, by Erran Carey.  
+Copyright, 2013, by Timothy Redaelli.  
+Copyright, 2013, by Jean-Christophe Cyr.  
+Copyright, 2013, by Keith Larrimore.  
+Copyright, 2014, by justfalter.  
+Copyright, 2014, by Zac Sprackett.  
+Copyright, 2014, by Mark Van de Vyver.  
+Copyright, 2014, by Peter M. Goldstein.  
+Copyright, 2014, by The Gitter Badger.  
+Copyright, 2014, by Chris Cunningham.  
+Copyright, 2015, by Michal Cichra.  
+Copyright, 2015, by Alexey Pisarenko.  
+Copyright, 2016, by John Bachir.  
+Copyright, 2016, by kaleforsale.  
+Copyright, 2018, by Rob Fors.  
+Copyright, 2020, by Olle Jonsson.  
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
