--- spec/rubydns/hosts.txt
@@ -1,2 +1,0 @@
-# A testing host:
-1.2.3.4 testing apples
