--- spec/spec_helper.rb
@@ -1,14 +1,0 @@
-
-require 'covered/rspec'
-require "bundler/setup"
-require "async"
-require "async/rspec"
-
-RSpec.configure do |config|
-	# Enable flags like --only-failures and --next-failure
-	config.example_status_persistence_file_path = ".rspec_status"
-
-	config.expect_with :rspec do |c|
-		c.syntax = :expect
-	end
-end
